@@ -15,8 +15,6 @@ class CURLRequest extends BaseConfig
      *
      * If true, all the options won't be reset between requests.
      * It may cause an error request with unnecessary headers.
-     *
-     * @var bool
      */
-    public $shareOptions = true;
+    public bool $shareOptions = false;
 }
