@@ -13,10 +13,8 @@ class Cookie extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * Set a cookie name prefix if you need to avoid collisions.
-     *
-     * @var string
      */
-    public $prefix = '';
+    public string $prefix = '';
 
     /**
      * --------------------------------------------------------------------------
@@ -37,10 +35,8 @@ class Cookie extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * Typically will be a forward slash.
-     *
-     * @var string
      */
-    public $path = '/';
+    public string $path = '/';
 
     /**
      * --------------------------------------------------------------------------
@@ -48,10 +44,8 @@ class Cookie extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * Set to `.your-domain.com` for site-wide cookies.
-     *
-     * @var string
      */
-    public $domain = '';
+    public string $domain = '';
 
     /**
      * --------------------------------------------------------------------------
@@ -59,10 +53,8 @@ class Cookie extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * Cookie will only be set if a secure HTTPS connection exists.
-     *
-     * @var bool
      */
-    public $secure = false;
+    public bool $secure = false;
 
     /**
      * --------------------------------------------------------------------------
@@ -70,10 +62,8 @@ class Cookie extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * Cookie will only be accessible via HTTP(S) (no JavaScript).
-     *
-     * @var bool
      */
-    public $httponly = true;
+    public bool $httponly = true;
 
     /**
      * --------------------------------------------------------------------------
@@ -95,9 +85,9 @@ class Cookie extends BaseConfig
      * (empty string) means default SameSite attribute set by browsers (`Lax`)
      * will be set on cookies. If set to `None`, `$secure` must also be set.
      *
-     * @var string
+     * @phpstan-var 'None'|'Lax'|'Strict'|''
      */
-    public $samesite = 'Lax';
+    public string $samesite = 'Lax';
 
     /**
      * --------------------------------------------------------------------------
@@ -110,10 +100,8 @@ class Cookie extends BaseConfig
      * If this is set to `true`, cookie names should be compliant of RFC 2616's
      * list of allowed characters.
      *
-     * @var bool
-     *
      * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie#attributes
      * @see https://tools.ietf.org/html/rfc2616#section-2.2
      */
-    public $raw = false;
+    public bool $raw = false;
 }
