@@ -6,165 +6,116 @@ use CodeIgniter\Config\BaseConfig;
 
 class Email extends BaseConfig
 {
-    /**
-     * @var string
-     */
-    public $fromEmail;
-
-    /**
-     * @var string
-     */
-    public $fromName;
-
-    /**
-     * @var string
-     */
-    public $recipients;
+    public string $fromEmail  = '';
+    public string $fromName   = '';
+    public string $recipients = '';
 
     /**
      * The "user agent"
-     *
-     * @var string
      */
-    public $userAgent = 'CodeIgniter';
+    public string $userAgent = 'CodeIgniter';
 
     /**
      * The mail sending protocol: mail, sendmail, smtp
-     *
-     * @var string
      */
-    public $protocol = 'mail';
+    public string $protocol = 'mail';
 
     /**
      * The server path to Sendmail.
-     *
-     * @var string
      */
-    public $mailPath = '/usr/sbin/sendmail';
+    public string $mailPath = '/usr/sbin/sendmail';
 
     /**
-     * SMTP Server Address
-     *
-     * @var string
+     * SMTP Server Hostname
      */
-    public $SMTPHost;
+    public string $SMTPHost = '';
 
     /**
      * SMTP Username
-     *
-     * @var string
      */
-    public $SMTPUser;
+    public string $SMTPUser = '';
 
     /**
      * SMTP Password
-     *
-     * @var string
      */
-    public $SMTPPass;
+    public string $SMTPPass = '';
 
     /**
      * SMTP Port
-     *
-     * @var int
      */
-    public $SMTPPort = 25;
+    public int $SMTPPort = 25;
 
     /**
      * SMTP Timeout (in seconds)
-     *
-     * @var int
      */
-    public $SMTPTimeout = 5;
+    public int $SMTPTimeout = 5;
 
     /**
      * Enable persistent SMTP connections
-     *
-     * @var bool
      */
-    public $SMTPKeepAlive = false;
+    public bool $SMTPKeepAlive = false;
 
     /**
-     * SMTP Encryption. Either tls or ssl
+     * SMTP Encryption.
      *
-     * @var string
+     * @var string '', 'tls' or 'ssl'. 'tls' will issue a STARTTLS command
+     *             to the server. 'ssl' means implicit SSL. Connection on port
+     *             465 should set this to ''.
      */
-    public $SMTPCrypto = 'tls';
+    public string $SMTPCrypto = 'tls';
 
     /**
      * Enable word-wrap
-     *
-     * @var bool
      */
-    public $wordWrap = true;
+    public bool $wordWrap = true;
 
     /**
      * Character count to wrap at
-     *
-     * @var int
      */
-    public $wrapChars = 76;
+    public int $wrapChars = 76;
 
     /**
      * Type of mail, either 'text' or 'html'
-     *
-     * @var string
      */
-    public $mailType = 'text';
+    public string $mailType = 'text';
 
     /**
      * Character set (utf-8, iso-8859-1, etc.)
-     *
-     * @var string
      */
-    public $charset = 'UTF-8';
+    public string $charset = 'UTF-8';
 
     /**
      * Whether to validate the email address
-     *
-     * @var bool
      */
-    public $validate = false;
+    public bool $validate = false;
 
     /**
      * Email Priority. 1 = highest. 5 = lowest. 3 = normal
-     *
-     * @var int
      */
-    public $priority = 3;
+    public int $priority = 3;
 
     /**
      * Newline character. (Use “\r\n” to comply with RFC 822)
-     *
-     * @var string
      */
-    public $CRLF = "\r\n";
+    public string $CRLF = "\r\n";
 
     /**
      * Newline character. (Use “\r\n” to comply with RFC 822)
-     *
-     * @var string
      */
-    public $newline = "\r\n";
+    public string $newline = "\r\n";
 
     /**
      * Enable BCC Batch Mode.
-     *
-     * @var bool
      */
-    public $BCCBatchMode = false;
+    public bool $BCCBatchMode = false;
 
     /**
      * Number of emails in each BCC batch
-     *
-     * @var int
      */
-    public $BCCBatchSize = 200;
+    public int $BCCBatchSize = 200;
 
     /**
      * Enable notify message from server
-     *
-     * @var bool
      */
-    public $DSN = false;
+    public bool $DSN = false;
 }
