@@ -20,10 +20,8 @@ class Encryption extends BaseConfig
      * If you use the Encryption class you must set an encryption key (seed).
      * You need to ensure it is long enough for the cipher and mode you plan to use.
      * See the user guide for more info.
-     *
-     * @var string
      */
-    public $key = '';
+    public string $key = '';
 
     /**
      * --------------------------------------------------------------------------
@@ -35,10 +33,8 @@ class Encryption extends BaseConfig
      * Available drivers:
      * - OpenSSL
      * - Sodium
-     *
-     * @var string
      */
-    public $driver = 'OpenSSL';
+    public string $driver = 'OpenSSL';
 
     /**
      * --------------------------------------------------------------------------
@@ -49,10 +45,8 @@ class Encryption extends BaseConfig
      * before it is encrypted. This value should be greater than zero.
      *
      * See the user guide for more information on padding.
-     *
-     * @var int
      */
-    public $blockSize = 16;
+    public int $blockSize = 16;
 
     /**
      * --------------------------------------------------------------------------
@@ -60,8 +54,39 @@ class Encryption extends BaseConfig
      * --------------------------------------------------------------------------
      *
      * HMAC digest to use, e.g. 'SHA512' or 'SHA256'. Default value is 'SHA512'.
-     *
-     * @var string
      */
-    public $digest = 'SHA512';
+    public string $digest = 'SHA512';
+
+    /**
+     * Whether the cipher-text should be raw. If set to false, then it will be base64 encoded.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to false for CI3 Encryption compatibility.
+     */
+    public bool $rawData = true;
+
+    /**
+     * Encryption key info.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to 'encryption' for CI3 Encryption compatibility.
+     */
+    public string $encryptKeyInfo = '';
+
+    /**
+     * Authentication key info.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to 'authentication' for CI3 Encryption compatibility.
+     */
+    public string $authKeyInfo = '';
+
+    /**
+     * Cipher to use.
+     * This setting is only used by OpenSSLHandler.
+     *
+     * Set to 'AES-128-CBC' to decrypt encrypted data that encrypted
+     * by CI3 Encryption default configuration.
+     */
+    public string $cipher = 'AES-256-CTR';
 }
