@@ -10,23 +10,28 @@ use CodeIgniter\Config\BaseConfig;
 class Feature extends BaseConfig
 {
     /**
-     * Enable multiple filters for a route or not.
-     *
-     * If you enable this:
-     *   - CodeIgniter\CodeIgniter::handleRequest() uses:
-     *     - CodeIgniter\Filters\Filters::enableFilters(), instead of enableFilter()
-     *   - CodeIgniter\CodeIgniter::tryToRouteIt() uses:
-     *     - CodeIgniter\Router\Router::getFilters(), instead of getFilter()
-     *   - CodeIgniter\Router\Router::handle() uses:
-     *     - property $filtersInfo, instead of $filterInfo
-     *     - CodeIgniter\Router\RouteCollection::getFiltersForRoute(), instead of getFilterForRoute()
-     *
-     * @var bool
+     * Use improved new auto routing instead of the legacy version.
      */
-    public $multipleFilters = false;
+    public bool $autoRoutesImproved = true;
 
     /**
-     * Use improved new auto routing instead of the default legacy version.
+     * Use filter execution order in 4.4 or before.
      */
-    public bool $autoRoutesImproved = false;
+    public bool $oldFilterOrder = false;
+
+    /**
+     * The behavior of `limit(0)` in Query Builder.
+     *
+     * If true, `limit(0)` returns all records. (the behavior of 4.4.x or before in version 4.x.)
+     * If false, `limit(0)` returns no records. (the behavior of 3.1.9 or later in version 3.x.)
+     */
+    public bool $limitZeroAsAll = true;
+
+    /**
+     * Use strict location negotiation.
+     *
+     * By default, the locale is selected based on a loose comparison of the language code (ISO 639-1)
+     * Enabling strict comparison will also consider the region code (ISO 3166-1 alpha-2).
+     */
+    public bool $strictLocaleNegotiation = false;
 }
