@@ -8,36 +8,35 @@ class Honeypot extends BaseConfig
 {
     /**
      * Makes Honeypot visible or not to human
-     *
-     * @var bool
      */
-    public $hidden = true;
+    public bool $hidden = true;
 
     /**
      * Honeypot Label Content
-     *
-     * @var string
      */
-    public $label = 'Fill This Field';
+    public string $label = 'Fill This Field';
 
     /**
      * Honeypot Field Name
-     *
-     * @var string
      */
-    public $name = 'honeypot';
+    public string $name = 'honeypot';
 
     /**
      * Honeypot HTML Template
-     *
-     * @var string
      */
-    public $template = '<label>{label}</label><input type="text" name="{name}" value=""/>';
+    public string $template = '<label>{label}</label><input type="text" name="{name}" value="">';
 
     /**
      * Honeypot container
      *
-     * @var string
+     * If you enabled CSP, you can remove `style="display:none"`.
      */
-    public $container = '<div style="display:none">{template}</div>';
+    public string $container = '<div style="display:none">{template}</div>';
+
+    /**
+     * The id attribute for Honeypot container tag
+     *
+     * Used when CSP is enabled.
+     */
+    public string $containerId = 'hpc';
 }
