@@ -10,25 +10,21 @@ class Images extends BaseConfig
 {
     /**
      * Default handler used if no other handler is specified.
-     *
-     * @var string
      */
-    public $defaultHandler = 'gd';
+    public string $defaultHandler = 'gd';
 
     /**
      * The path to the image library.
      * Required for ImageMagick, GraphicsMagick, or NetPBM.
-     *
-     * @var string
      */
-    public $libraryPath = '/usr/local/bin/convert';
+    public string $libraryPath = '/usr/local/bin/convert';
 
     /**
      * The available handler classes.
      *
      * @var array<string, string>
      */
-    public $handlers = [
+    public array $handlers = [
         'gd'      => GDHandler::class,
         'imagick' => ImageMagickHandler::class,
     ];
