@@ -2,8 +2,9 @@
 
 namespace Config;
 
-use CodeIgniter\Config\BaseConfig;
-use Kint\Renderer\Renderer;
+use Kint\Parser\ConstructablePluginInterface;
+use Kint\Renderer\Rich\TabPluginInterface;
+use Kint\Renderer\Rich\ValuePluginInterface;
 
 /**
  * --------------------------------------------------------------------------
@@ -15,7 +16,7 @@ use Kint\Renderer\Renderer;
  *
  * @see https://kint-php.github.io/kint/ for details on these settings.
  */
-class Kint extends BaseConfig
+class Kint
 {
     /*
     |--------------------------------------------------------------------------
@@ -23,20 +24,31 @@ class Kint extends BaseConfig
     |--------------------------------------------------------------------------
     */
 
+    /**
+     * @var list<class-string<ConstructablePluginInterface>|ConstructablePluginInterface>|null
+     */
     public $plugins;
-    public $maxDepth          = 6;
-    public $displayCalledFrom = true;
-    public $expanded          = false;
+
+    public int $maxDepth           = 6;
+    public bool $displayCalledFrom = true;
+    public bool $expanded          = false;
 
     /*
     |--------------------------------------------------------------------------
     | RichRenderer Settings
     |--------------------------------------------------------------------------
     */
-    public $richTheme  = 'aante-light.css';
-    public $richFolder = false;
-    public $richSort   = Renderer::SORT_FULL;
+    public string $richTheme = 'aante-light.css';
+    public bool $richFolder  = false;
+
+    /**
+     * @var array<string, class-string<ValuePluginInterface>>|null
+     */
     public $richObjectPlugins;
+
+    /**
+     * @var array<string, class-string<TabPluginInterface>>|null
+     */
     public $richTabPlugins;
 
     /*
@@ -44,8 +56,8 @@ class Kint extends BaseConfig
     | CLI Settings
     |--------------------------------------------------------------------------
     */
-    public $cliColors      = true;
-    public $cliForceUTF8   = false;
-    public $cliDetectWidth = true;
-    public $cliMinWidth    = 40;
+    public bool $cliColors      = true;
+    public bool $cliForceUTF8   = false;
+    public bool $cliDetectWidth = true;
+    public int $cliMinWidth     = 40;
 }
