@@ -15,10 +15,8 @@ class Migrations extends BaseConfig
      *
      * You should enable migrations whenever you intend to do a schema migration
      * and disable it back when you're done.
-     *
-     * @var bool
      */
-    public $enabled = true;
+    public bool $enabled = true;
 
     /**
      * --------------------------------------------------------------------------
@@ -27,13 +25,9 @@ class Migrations extends BaseConfig
      *
      * This is the name of the table that will store the current migrations state.
      * When migrations runs it will store in a database table which migration
-     * level the system is at. It then compares the migration level in this
-     * table to the $config['migration_version'] if they are not the same it
-     * will migrate up. This must be set.
-     *
-     * @var string
+     * files have already been run.
      */
-    public $table = 'migrations';
+    public string $table = 'migrations';
 
     /**
      * --------------------------------------------------------------------------
@@ -42,14 +36,15 @@ class Migrations extends BaseConfig
      *
      * This is the format that will be used when creating new migrations
      * using the CLI command:
-     *   > php spark migrate:create
+     *   > php spark make:migration
      *
-     * Typical formats:
+     * NOTE: if you set an unsupported format, migration runner will not find
+     *       your migration files.
+     *
+     * Supported formats:
      * - YmdHis_
      * - Y-m-d-His_
      * - Y_m_d_His_
-     *
-     * @var string
      */
-    public $timestampFormat = 'Y-m-d-His_';
+    public string $timestampFormat = 'Y-m-d-His_';
 }
