@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title><?= lang('Errors.pageNotFound') ?></title>
+    <title><?= lang('Errors.badRequest') ?></title>
 
     <style>
         div.logo {
@@ -69,16 +69,16 @@
     </style>
 </head>
 <body>
-    <div class="wrap">
-        <h1>404</h1>
+<div class="wrap">
+    <h1>400</h1>
 
-        <p>
-            <?php if (ENVIRONMENT !== 'production') : ?>
-                <?= nl2br(esc($message)) ?>
-            <?php else : ?>
-                <?= lang('Errors.sorryCannotFind') ?>
-            <?php endif; ?>
-        </p>
-    </div>
+    <p>
+        <?php if (ENVIRONMENT !== 'production') : ?>
+            <?= nl2br(esc($message)) ?>
+        <?php else : ?>
+            <?= lang('Errors.sorryBadRequest') ?>
+        <?php endif; ?>
+    </p>
+</div>
 </body>
 </html>
