@@ -4,9 +4,9 @@
     <meta charset="UTF-8">
     <meta name="robots" content="noindex">
 
-    <title>Whoops!</title>
+    <title><?= lang('Errors.whoops') ?></title>
 
-    <style type="text/css">
+    <style>
         <?= preg_replace('#[\r\n\t ]+#', ' ', file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . 'debug.css')) ?>
     </style>
 </head>
@@ -14,9 +14,9 @@
 
     <div class="container text-center">
 
-        <h1 class="headline">Whoops!</h1>
+        <h1 class="headline"><?= lang('Errors.whoops') ?></h1>
 
-        <p class="lead">We seem to have hit a snag. Please try again later...</p>
+        <p class="lead"><?= lang('Errors.weHitASnag') ?></p>
 
     </div>
 
