@@ -1,16 +1,15 @@
 <?php
 
 use CodeIgniter\Test\CIUnitTestCase;
-use Config\Services;
 
 /**
  * @internal
  */
 final class ExampleSessionTest extends CIUnitTestCase
 {
-    public function testSessionSimple()
+    public function testSessionSimple(): void
     {
-        $session = Services::session();
+        $session = service('session');
 
         $session->set('logged_in', 123);
         $this->assertSame(123, $session->get('logged_in'));
