@@ -2,7 +2,6 @@
 
 use CodeIgniter\Test\CIUnitTestCase;
 use Config\App;
-use Config\Services;
 use Tests\Support\Libraries\ConfigReader;
 
 /**
@@ -10,14 +9,14 @@ use Tests\Support\Libraries\ConfigReader;
  */
 final class HealthTest extends CIUnitTestCase
 {
-    public function testIsDefinedAppPath()
+    public function testIsDefinedAppPath(): void
     {
         $this->assertTrue(defined('APPPATH'));
     }
 
-    public function testBaseUrlHasBeenSet()
+    public function testBaseUrlHasBeenSet(): void
     {
-        $validation = Services::validation();
+        $validation = service('validation');
 
         $env = false;
 
@@ -33,7 +32,7 @@ final class HealthTest extends CIUnitTestCase
             $config = new App();
             $this->assertTrue(
                 $validation->check($config->baseURL, 'valid_url'),
-                'baseURL "' . $config->baseURL . '" in .env is not valid URL'
+                'baseURL "' . $config->baseURL . '" in .env is not valid URL',
             );
         }
 
@@ -44,7 +43,7 @@ final class HealthTest extends CIUnitTestCase
         // BaseURL in app/Config/App.php is a valid URL?
         $this->assertTrue(
             $validation->check($reader->baseURL, 'valid_url'),
-            'baseURL "' . $reader->baseURL . '" in app/Config/App.php is not valid URL'
+            'baseURL "' . $reader->baseURL . '" in app/Config/App.php is not valid URL',
         );
     }
 }
